-work on handle all the projects and modifying all files of a  du
